Handling checks for category not found

+++ src/main/java/com/ecommerce/sb_ecom/model/Category.java
@@ -2,19 +2,19 @@ package com.ecommerce.sb_ecom.model;
 
 public class Category
 {
-    private int categoryId;
+    private Long categoryId;
     private String categoryName;
 
-    public Category(int categoryId, String categoryName) {
+    public Category(Long categoryId, String categoryName) {
         this.categoryId = categoryId;
         this.categoryName = categoryName;
     }
 
-    public int getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(long categoryId) {
         this.categoryId = categoryId;
     }
